app/api/semantic_search/requirements.txt
fastapi
uvicorn
pydantic
pymongo
sentence-transformers
numpy
pinecone
python-dotenv
